Added swagger and update folder structure
