+++ src/Docs/Admin/Controllers/Customers/GroupController.php
@@ -1,0 +1,300 @@
+<?php
+
+namespace Webkul\RestApi\Docs\Admin\Controllers\Customers;
+
+class GroupController
+{
+    /**
+     * @OA\Get(
+     *      path="/api/v1/admin/customers/groups",
+     *      operationId="getCustomerGroups",
+     *      tags={"CustomerGroups"},
+     *      summary="Get admin customer group list",
+     *      description="Returns customer group list, if you want to retrieve all customer groups at once pass pagination=0 otherwise ignore this parameter",
+     *      security={ {"sanctum_admin": {} }},
+     *
+     *      @OA\Parameter(
+     *          name="id",
+     *          description="Customer Group Id",
+     *          required=false,
+     *          in="query",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\Parameter(
+     *          name="sort",
+     *          description="Sort column",
+     *          example="id",
+     *          required=false,
+     *          in="query",
+     *
+     *          @OA\Schema(
+     *              type="string"
+     *          )
+     *      ),
+     *
+     *      @OA\Parameter(
+     *          name="order",
+     *          description="Sort order",
+     *          required=false,
+     *          in="query",
+     *
+     *          @OA\Schema(
+     *              type="string",
+     *              enum={"desc", "asc"}
+     *          )
+     *      ),
+     *
+     *      @OA\Parameter(
+     *          name="page",
+     *          description="Page number",
+     *          required=false,
+     *          in="query",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\Parameter(
+     *          name="limit",
+     *          description="Limit",
+     *          in="query",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Property(
+     *                  property="data",
+     *                  type="array",
+     *
+     *                  @OA\Items(ref="#/components/schemas/Group")
+     *              ),
+     *
+     *              @OA\Property(
+     *                  property="meta",
+     *                  ref="#/components/schemas/Pagination"
+     *              )
+     *          )
+     *      )
+     * )
+     */
+    public function list()
+    {
+    }
+
+    /**
+     * @OA\Get(
+     *      path="/api/v1/admin/customers/groups/{id}",
+     *      operationId="getCustomerGroup",
+     *      tags={"CustomerGroups"},
+     *      summary="Get admin customer group detail",
+     *      description="Returns customer group detail",
+     *      security={ {"sanctum_admin": {} }},
+     *
+     *      @OA\Parameter(
+     *          name="id",
+     *          description="Customer Group ID",
+     *          required=true,
+     *          in="path",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Property(
+     *                  property="data",
+     *                  type="object",
+     *                  ref="#/components/schemas/Group"
+     *              )
+     *          )
+     *      )
+     * )
+     */
+    public function get()
+    {
+    }
+
+    /**
+     * @OA\Post(
+     *      path="/api/v1/admin/customers/groups",
+     *      operationId="storeCustomerGroup",
+     *      tags={"CustomerGroups"},
+     *      summary="Store the customer group",
+     *      description="Store the customer group",
+     *      security={ {"sanctum_admin": {} }},
+     *
+     *      @OA\RequestBody(
+     *
+     *          @OA\MediaType(
+     *              mediaType="multipart/form-data",
+     *
+     *              @OA\Schema(
+     *
+     *                  @OA\Property(
+     *                      property="code",
+     *                      type="string",
+     *                      description="Customer group's code `Unique`, `No-whitespace`, `No-underscore`",
+     *                      example="vip"
+     *                  ),
+     *                  @OA\Property(
+     *                      property="name",
+     *                      type="string",
+     *                      description="Customer group's name",
+     *                      example="VIP Group"
+     *                  ),
+     *                  required={"code", "name"}
+     *              )
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Property(property="message", type="string", example="Customer group created successfully."),
+     *              @OA\Property(property="data", type="object", ref="#/components/schemas/Group")
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      )
+     * )
+     */
+    public function store()
+    {
+    }
+
+    /**
+     * @OA\Put(
+     *      path="/api/v1/admin/customers/groups/{id}",
+     *      operationId="updateCustomerGroup",
+     *      tags={"CustomerGroups"},
+     *      summary="Update customer group",
+     *      description="Update customer group",
+     *      security={ {"sanctum_admin": {} }},
+     *
+     *      @OA\Parameter(
+     *          name="id",
+     *          description="Customer Group ID",
+     *          required=true,
+     *          in="path",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\RequestBody(
+     *
+     *          @OA\MediaType(
+     *              mediaType="application/json",
+     *
+     *              @OA\Schema(
+     *
+     *                  @OA\Property(
+     *                      property="code",
+     *                      type="string",
+     *                      description="Customer group's code `Unique`, `No-whitespace`, `No-underscore`",
+     *                      example="vip"
+     *                  ),
+     *                  @OA\Property(
+     *                      property="name",
+     *                      type="string",
+     *                      description="Customer group's name",
+     *                      example="VIP Group"
+     *                  ),
+     *                  required={"code", "name"}
+     *              )
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Property(
+     *                  property="message",
+     *                  type="string",
+     *                  example="Customer group updated successfully."),
+     *              @OA\Property(
+     *                  property="data",
+     *                  type="object",
+     *                  ref="#/components/schemas/Group"
+     *              )
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=401,
+     *          description="Unauthenticated",
+     *      )
+     * )
+     */
+    public function update()
+    {
+    }
+
+    /**
+     * @OA\Delete(
+     *      path="/api/v1/admin/customers/groups/{id}",
+     *      operationId="deleteCustomerGroup",
+     *      tags={"CustomerGroups"},
+     *      summary="Delete customer group by id",
+     *      description="Delete customer group by id",
+     *      security={ {"sanctum_admin": {} }},
+     *
+     *      @OA\Parameter(
+     *          name="id",
+     *          description="Customer Group ID",
+     *          required=true,
+     *          in="path",
+     *
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *      ),
+     *
+     *      @OA\Response(
+     *          response=200,
+     *          description="Successful operation",
+     *
+     *          @OA\JsonContent(
+     *
+     *              @OA\Property(
+     *                  property="message",
+     *                  type="string",
+     *                  example="Customer group deleted successfully."),
+     *              )
+     *          )
+     *      )
+     * )
+     */
+    public function destroy()
+    {
+    }
+}
