+++ src/Docs/Admin/Controllers/Marketing/Communications/EventController.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace Webkul\RestApi\Docs\Admin\Controllers\Marketing;
+namespace Webkul\RestApi\Docs\Admin\Controllers\Marketing\Communications;
 
-class EmailTemplateController
+class EventController
 {
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/promotions/email-templates",
-     *      operationId="getEmailTemplates",
-     *      tags={"EmailTemplates"},
-     *      summary="Get admin email template list",
-     *      description="Returns email template list, if you want to retrieve all email templates at once pass pagination=0 otherwise ignore this parameter",
+     *      path="/api/v1/admin/promotions/events",
+     *      operationId="getEvents",
+     *      tags={"Events"},
+     *      summary="Get admin event list",
+     *      description="Returns event list, if you want to retrieve all events at once pass pagination=0 otherwise ignore this parameter",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Email Template Id",
+     *          description="Event Id",
      *          required=false,
      *          in="query",
      *
@@ -79,7 +79,7 @@ class EmailTemplateController
      *                  property="data",
      *                  type="array",
      *
-     *                  @OA\Items(ref="#/components/schemas/EmailTemplate")
+     *                  @OA\Items(ref="#/components/schemas/Event")
      *              ),
      *
      *              @OA\Property(
@@ -96,16 +96,16 @@ class EmailTemplateController
 
     /**
      * @OA\Get(
-     *      path="/api/v1/admin/promotions/email-templates/{id}",
-     *      operationId="getEmailTemplate",
-     *      tags={"EmailTemplates"},
-     *      summary="Get admin email template detail",
-     *      description="Returns email template detail",
+     *      path="/api/v1/admin/promotions/events/{id}",
+     *      operationId="getEvent",
+     *      tags={"Events"},
+     *      summary="Get admin event detail",
+     *      description="Returns event detail",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Email Template ID",
+     *          description="Event ID",
      *          required=true,
      *          in="path",
      *
@@ -123,7 +123,7 @@ class EmailTemplateController
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/EmailTemplate"
+     *                  ref="#/components/schemas/Event"
      *              )
      *          )
      *      )
@@ -135,11 +135,11 @@ class EmailTemplateController
 
     /**
      * @OA\Post(
-     *      path="/api/v1/admin/promotions/email-templates",
-     *      operationId="storeEmailTemplate",
-     *      tags={"EmailTemplates"},
-     *      summary="Store the email template",
-     *      description="Store the email template",
+     *      path="/api/v1/admin/promotions/events",
+     *      operationId="storeEvent",
+     *      tags={"Events"},
+     *      summary="Store the event",
+     *      description="Store the event",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\RequestBody(
@@ -152,23 +152,23 @@ class EmailTemplateController
      *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      description="Email template name",
-     *                      example="Festival Offer"
+     *                      description="Event name",
+     *                      example="Birthday Offer"
      *                  ),
      *                  @OA\Property(
-     *                      property="content",
+     *                      property="description",
      *                      type="string",
      *                      description="Email template content",
      *                      example="What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
      *                  ),
      *                  @OA\Property(
-     *                      property="status",
+     *                      property="date",
+     *                      description="Event date",
+     *                      format="date",
      *                      type="string",
-     *                      description="Email template status",
-     *                      example="inactive",
-     *                      enum={"active", "inactive", "draft"}
+     *                      example="2024-01-27"
      *                  ),
-     *                  required={"name", "content"}
+     *                  required={"name", "description", "date"}
      *              )
      *          )
      *      ),
@@ -179,8 +179,8 @@ class EmailTemplateController
      *
      *          @OA\JsonContent(
      *
-     *              @OA\Property(property="message", type="string", example="Email template created successfully."),
-     *              @OA\Property(property="data", type="object", ref="#/components/schemas/EmailTemplate")
+     *              @OA\Property(property="message", type="string", example="Event created successfully."),
+     *              @OA\Property(property="data", type="object", ref="#/components/schemas/Event")
      *          )
      *      ),
      *
@@ -196,16 +196,16 @@ class EmailTemplateController
 
     /**
      * @OA\Put(
-     *      path="/api/v1/admin/promotions/email-templates/{id}",
-     *      operationId="updateEmailTemplate",
-     *      tags={"EmailTemplates"},
-     *      summary="Update email template",
-     *      description="Update email template",
+     *      path="/api/v1/admin/promotions/events/{id}",
+     *      operationId="updateEvent",
+     *      tags={"Events"},
+     *      summary="Update event",
+     *      description="Update event",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Email Template ID",
+     *          description="Event ID",
      *          required=true,
      *          in="path",
      *
@@ -224,23 +224,23 @@ class EmailTemplateController
      *                  @OA\Property(
      *                      property="name",
      *                      type="string",
-     *                      description="Email template name",
-     *                      example="Festival Offer"
+     *                      description="Event name",
+     *                      example="Anniversary Offer"
      *                  ),
      *                  @OA\Property(
-     *                      property="content",
+     *                      property="description",
      *                      type="string",
      *                      description="Email template content",
      *                      example="What is Lorem Ipsum? Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum."
      *                  ),
      *                  @OA\Property(
-     *                      property="status",
+     *                      property="date",
+     *                      description="Event date",
+     *                      format="date",
      *                      type="string",
-     *                      description="Email template status",
-     *                      example="inactive",
-     *                      enum={"active", "inactive", "draft"}
+     *                      example="2024-01-27"
      *                  ),
-     *                  required={"name", "content"}
+     *                  required={"name", "description", "date"}
      *              )
      *          )
      *      ),
@@ -254,11 +254,11 @@ class EmailTemplateController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="Email template updated successfully."),
+     *                  example="Event updated successfully."),
      *              @OA\Property(
      *                  property="data",
      *                  type="object",
-     *                  ref="#/components/schemas/EmailTemplate"
+     *                  ref="#/components/schemas/Event"
      *              )
      *          )
      *      ),
@@ -275,16 +275,16 @@ class EmailTemplateController
 
     /**
      * @OA\Delete(
-     *      path="/api/v1/admin/promotions/email-templates/{id}",
-     *      operationId="deleteEmailTemplate",
-     *      tags={"EmailTemplates"},
-     *      summary="Delete email template by id",
-     *      description="Delete email template by id",
+     *      path="/api/v1/admin/promotions/events/{id}",
+     *      operationId="deleteEvent",
+     *      tags={"Events"},
+     *      summary="Delete event by id",
+     *      description="Delete event by id",
      *      security={ {"sanctum_admin": {} }},
      *
      *      @OA\Parameter(
      *          name="id",
-     *          description="Email Template ID",
+     *          description="Event ID",
      *          required=true,
      *          in="path",
      *
@@ -302,7 +302,7 @@ class EmailTemplateController
      *              @OA\Property(
      *                  property="message",
      *                  type="string",
-     *                  example="Email template deleted successfully."),
+     *                  example="Event deleted successfully."),
      *              )
      *          )
      *      )
